Insurance_Document_Query_System_Using_RAG/requirements.txt
streamlit
langchain_core
langchain_community
langchain_ollama
faiss-cpu 
pytesseract
llama-index
llama_index-llms-ollama
python-dotenv
llama-parse
langchain_text_splitters